script rename
